Fixed bug where Harris always returns the same class_year.
This bug only occurs with the production server (as a result of an
optimization?). In order to successfully authenticate, curl has to
enable cookies. Once that cookie exists, any request made to
Harris using that cookie will return the same class_year. All
other fields will change appropriately. The fix is to just use a
different cookie for every login.